+++ wiimote/include/wiimote/wiimote_controller.hpp
@@ -185,7 +185,7 @@ private:
   uint8_t classic_stick_right_min_[2];
 
   const int IGNORE_DATA_POINTS_ = 100;  // throw away the first few data points
-  const int COVARIANCE_DATA_POINTS_ = 100;
+  const size_t COVARIANCE_DATA_POINTS_ = 100;
   StatVector3d linear_acceleration_stat_;
   StatVector3d angular_velocity_stat_;
   boost::array<double, 9> linear_acceleration_covariance_;
